Update KafkaReader.java
Camus has a issues and failing while comparing the Checksum of the message once partition ID is used to persists the message. The issues is with the way Camus read the messages from Kafka using EtlRecordReader.java and KafkaReader.java class. If the message is persisted with partition ID, Kafka basically persists the partition ID along with the message. But when Camus compute the checksum, it only takes the payload, not the complete message (i.e. message with Partition Key). Thus the checksum is failing.