Added sort to guarantee ordering of executions
When Camus is deleting old executions and finding the latest, the
ordering of executions was not guaranteed. (i.e. Camus may find a
previous execution that isn't the latest). This has been an issue when
testing Camus using the local file system or under certain Hadoop
distributions. The listing of execution files is now sorted into 'last
modified' order.